Cập Nhật Lần Cuối

+++ ShopBanHangHS/Areas/Admin/Controllers/DanhMucAdminController.cs
@@ -4,7 +4,9 @@ using Microsoft.EntityFrameworkCore;
 using ShopBanHangHS.Areas.Admin.Data;
 using ShopBanHangHS.Areas.Admin.Models;
 using ShopBanHangHS.Data;
+using ShopBanHangHS.Help;
 using System;
+using System.Collections.Generic;
 using System.Linq;
 
 namespace ShopBanHangHS.Areas.Admin.Controllers
@@ -16,31 +18,81 @@ namespace ShopBanHangHS.Areas.Admin.Controllers
         public DanhMucAdminController (DBContextAdmin _db)
         {
             db = _db;
-        }       
+        }
+        public List<User> Userss
+        {
+            get
+            {
+                var data = HttpContext.Session.Get<List<User>>("Tk");
+                if (data == null)
+                {
+                    data = new List<User>();
+
+                }
+                return data;
+            }
+        }
         public IActionResult Index()
         {
-            var check = db.DanhMucs.ToList();
-            return View(check);
+            var use = Userss;
+            var checkUs = use.FirstOrDefault();
+            if (checkUs == null)
+            {
+                return Redirect("/Login/Index");
+            }
+            else
+            {
+                var check = db.DanhMucs.ToList();
+                return View(check);
+            }
+            
         }
         public IActionResult Create()
         {
-            DanhMuc danhMuc = new DanhMuc();
-            return View(danhMuc);
+            var use = Userss;
+            var checkUs = use.FirstOrDefault();
+            if (checkUs == null)
+            {
+                return Redirect("/Login/Index");
+            }
+            else
+            {
+                DanhMuc danhMuc = new DanhMuc();
+                return View(danhMuc);
+            }
         }
 
         [HttpPost]
         public IActionResult Create(DanhMuc dm)
         {
+            var use = Userss;
+            var checkUs = use.FirstOrDefault();
+            if (checkUs == null)
+            {
+                return Redirect("/Login/Index");
+            }
+            else
+            {
                 db.Attach(dm);
                 db.DanhMucs.Add(dm);
-                db.SaveChanges();           
-                return Redirect("/Admin/DanhMucAdmin/Index");    
+                db.SaveChanges();
+                return Redirect("/Admin/DanhMucAdmin/Index");
+            }
         }
         [Route("Admin/[controller]/[action]")]
         public IActionResult Delete(int id)
         {
-            var check = db.DanhMucs.Where(x => x.maDanhMuc == id).FirstOrDefault();
-            return View(check);
+            var use = Userss;
+            var checkUs = use.FirstOrDefault();
+            if (checkUs == null)
+            {
+                return Redirect("/Login/Index");
+            }
+            else
+            {
+                var check = db.DanhMucs.Where(x => x.maDanhMuc == id).FirstOrDefault();
+                return View(check);
+            }
         }
         [Route("Admin/[controller]/[action]")]
         [HttpPost]
@@ -48,23 +100,49 @@ namespace ShopBanHangHS.Areas.Admin.Controllers
         {
             // db.Attach(bietThu);
             // db.Entry(id).State = EntityState.Deleted;
-            var check = db.DanhMucs.SingleOrDefault(s => s.maDanhMuc == id);
-            db.DanhMucs.Remove(check);
-            db.SaveChanges();
-            return Redirect("/Admin/DanhMucAdmin/Index");
+            var use = Userss;
+            var checkUs = use.FirstOrDefault();
+            if (checkUs == null)
+            {
+                return Redirect("/Login/Index");
+            }
+            else
+            {
+                var check = db.DanhMucs.SingleOrDefault(s => s.maDanhMuc == id);
+                db.DanhMucs.Remove(check);
+                db.SaveChanges();
+                return Redirect("/Admin/DanhMucAdmin/Index");
+            }
         }
         [Route("Admin/[controller]/[action]")]
         public ActionResult Details(int id)
-        { var check  = db.DanhMucs.SingleOrDefault(s=>s.maDanhMuc ==id);
-
-            return View(check);
+        {
+            var use = Userss;
+            var checkUs = use.FirstOrDefault();
+            if (checkUs == null)
+            {
+                return Redirect("/Login/Index");
+            }
+            else
+            {
+                var check = db.DanhMucs.SingleOrDefault(s => s.maDanhMuc == id);
+                return View(check);
+            }
         }
         [Route("Admin/[controller]/[action]")]
         public ActionResult Edit(int id)
         {
-            var objds = db.DanhMucs.Where(x => x.maDanhMuc == id).FirstOrDefault();
-            return View(objds);
-          
+            var use = Userss;
+            var checkUs = use.FirstOrDefault();
+            if (checkUs == null)
+            {
+                return Redirect("/Login/Index");
+            }
+            else
+            {
+                var objds = db.DanhMucs.Where(x => x.maDanhMuc == id).FirstOrDefault();
+                return View(objds);
+            }
         }
         // POST: SanPhamAdminController/Edit/5
         [Route("Admin/[controller]/[action]")]
@@ -74,10 +152,19 @@ namespace ShopBanHangHS.Areas.Admin.Controllers
         {
             try
             {
-                 db.Entry(danhMuc).State= EntityState.Modified;
+                var use = Userss;
+                var checkUs = use.FirstOrDefault();
+                if (checkUs == null)
+                {
+                    return Redirect("/Login/Index");
+                }
+                else
+                {
+                    db.Entry(danhMuc).State = EntityState.Modified;
                     db.SaveChanges();
-                
-                return Redirect("/Admin/DanhMucAdmin/Index");
+
+                    return Redirect("/Admin/DanhMucAdmin/Index");
+                }
             }
             catch
             {
